extract schema fs and theme utilities
