--- packages/platform-core/src/createShop/utils/index.ts
@@ -1,3 +1,0 @@
-// packages/platform-core/createShop/utils/index.ts
-export { copyTemplate } from "./copyTemplate";
-export { loadBaseTokens } from "./loadBaseTokens";
